Tons of Changes
User page is up with user info.
User avatars. Uer permissions and such are up. Most things are good when it comes to logins and workers and stuff.

Working on making the schedule rn.  Made the get_monday function which returns the monday of the week associated with the day you put in. Also a create_days function that creates a list of date() objects .
so:
>>func('12/17/19')
[date1, date2, etc]