Trying to add React
So i want to manipulate the dom and make the project more 1-pagey, so im trying to make a flask api and use react with it.

Currently flask can send a json object to react and you can see the object in the console but I havent learned how to use the data. It seems the data will be put in state and we can use it there.